pyproject.toml
[tool.poetry]
name = "sibdev-test"
version = "0.1.0"
description = ""
authors = ["Dair Sheishenaliev <dokadair@gmail.com>"]
readme = "README.md"
packages = [{include = "sibdev_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.4"
djangorestframework-simplejwt = "^5.3.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"
pytest-html = "^4.0.0"
snapshottest = "^0.6.0"
ipython = "^8.15.0"
celery = "^5.3.4"
requests = "^2.31.0"
psycopg2 = "^2.9.7"
redis = "^5.0.0"
django-celery-beat = "^2.5.0"
flake8 = "^6.1.0"
pytest-redis = "^3.0.2"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
